+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.3.0] - 2016-09-28
 - [#201][PR201]: Subscribe method is now exposed on the client to allow subscription of callback URL's
 
 ### Fixed
@@ -66,7 +68,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial SemVer release.
 
-[Unreleased]: https://github.com/gambol99/go-marathon/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/gambol99/go-marathon/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/gambol99/go-marathon/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/gambol99/go-marathon/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/gambol99/go-marathon/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/gambol99/go-marathon/compare/v0.0.1...v0.1.0